We expect the values to return floats instead of integers

+++ custom_components/nhc2/entities/thermoswitchx_multisensor_ambient_temperature.py
@@ -32,7 +32,7 @@ class Nhc2ThermoswitchxMultisensorAmbientTemperatureEntity(SensorEntity):
         return 'Ambient Temperature'
 
     @property
-    def native_value(self) -> int:
+    def native_value(self) -> float:
         return self._device.ambient_temperature
 
     def on_change(self):
